Refactor: controller: Rename controld_lrm.h to controld_execd.h
